FM2-121: Remove condition resource implementation for openmrs platform v2.0 -2.1
