--- api-2.0/src/test/java/org/openmrs/module/fhir2/TestFhirSpringConfiguration.java
@@ -1,20 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public License,
- * v. 2.0. If a copy of the MPL was not distributed with this file, You can
- * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
- * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
- *
- * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
- * graphic logo is a trademark of OpenMRS Inc.
- */
-package org.openmrs.module.fhir2;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.ImportResource;
-
-@Configuration
-@ImportResource({ "classpath:applicationContext-service.xml", "classpath:testApplicationContext.xml",
-        "classpath*:moduleApplicationContext.xml" })
-public class TestFhirSpringConfiguration {
-	
-}
