--- api-2.0/src/test/resources/Condition.hbm.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0"?>
-<!--
-    This Source Code Form is subject to the terms of the Mozilla Public License,
-    v. 2.0. If a copy of the MPL was not distributed with this file, You can
-    obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
-    the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
-
-    Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
-    graphic logo is a trademark of OpenMRS Inc.
--->
-<!DOCTYPE hibernate-mapping PUBLIC
-        "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
-
-        "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
-<hibernate-mapping auto-import="false">
-    <!-- Plural table name ("conditions") because "condition" is a MySQL reserved word -->
-    <class name="org.openmrs.module.emrapi.conditionslist.Condition" table="conditions">
-        <id name="conditionId" type="int" column="condition_id" unsaved-value="0">
-            <generator class="native">
-                <param name="sequence">conditions_condition_id_seq</param>
-            </generator>
-        </id>
-
-        <discriminator column="condition_id" insert="false"/>
-
-        <property name="uuid" type="java.lang.String"
-                  column="uuid" length="38" unique="true"/>
-        <property name="status" column="status" length="255" not-null="true">
-            <type name="org.hibernate.type.EnumType">
-                <param name="enumClass">org.openmrs.module.emrapi.conditionslist.Condition$Status</param>
-                <param name="type">12</param>
-            </type>
-        </property>
-        <property name="conditionNonCoded" type="java.lang.String" column="condition_non_coded" length="1024"/>
-        <property name="onsetDate" type="java.util.Date" column="onset_date"/>
-        <property name="additionalDetail" type="java.lang.String" column="additional_detail" length="1024"/>
-        <property name="endDate" type="java.util.Date" column="end_date"/>
-        <property name="dateCreated" type="java.util.Date" column="date_created" not-null="true"/>
-        <property name="voided" type="java.lang.Boolean" column="voided"/>
-        <property name="dateVoided" type="java.util.Date" column="date_voided"/>
-        <property name="voidReason" type="java.lang.String" column="void_reason" length="255"/>
-
-        <many-to-one name="previousCondition" class="org.openmrs.module.emrapi.conditionslist.Condition"
-                     column="previous_condition_id"/>
-        <many-to-one name="patient" column="patient_id" class="org.openmrs.Patient" not-null="true"/>
-        <many-to-one name="concept" column="concept_id" class="org.openmrs.Concept" not-null="true"/>
-        <many-to-one name="endReason" column="end_reason" class="org.openmrs.Concept"/>
-        <many-to-one name="creator" class="org.openmrs.User" not-null="true"/>
-        <many-to-one name="voidedBy" class="org.openmrs.User" column="voided_by"/>
-    </class>
-</hibernate-mapping>
