+++ omod/pom.xml
@@ -20,11 +20,6 @@
         </dependency>
 		<dependency>
 			<groupId>${project.parent.groupId}</groupId>
-			<artifactId>${project.parent.artifactId}-api-2.0</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>${project.parent.groupId}</groupId>
 			<artifactId>${project.parent.artifactId}-api-2.1</artifactId>
 			<version>${project.parent.version}</version>
 		</dependency>
